// src/database/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const TokenStatus = {
  ACTIVATE: 'ACTIVATE',
  DEACTIVATE: 'DEACTIVATE',
  REVOKED: 'REVOKED',
  TERMINATED: 'TERMINATED'
} as const

export type TokenStatus = (typeof TokenStatus)[keyof typeof TokenStatus]